backend/requirements.txt
# Backend Core
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
python-multipart==0.0.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task Queue & Cache
redis==5.0.1
celery==5.3.6

# AI / ML / NLP
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.1.post1
spacy==3.7.4
transformers==4.38.2
torch==2.2.1

# File Processing
pymupdf==1.23.26
pdfminer.six==20231228

# Real-time
websockets==12.0
python-socketio==5.11.1

# Testing & Dev
pytest==8.0.2
black==24.2.0
isort==5.13.2
httpx==0.27.0
